Generalize the code printing of And, Or, Not.
Move the printing of And, Or, and Not up in the class hierarchy to
the CodePrinter class and make it customizable by overriding a dict.
