+++ sympy/printing/ccode.py
@@ -15,7 +15,6 @@ from sympy.core import S, C
 from sympy.core.compatibility import string_types
 from sympy.printing.codeprinter import CodePrinter
 from sympy.printing.precedence import precedence
-from sympy.core.compatibility import default_sort_key
 
 # dictionary mapping sympy function to (argument_conditions, C_function).
 # Used in CCodePrinter._print_Function(self)
@@ -176,20 +175,6 @@ class CCodePrinter(CodePrinter):
                             self._print(expr.args[-1].expr)))
         code = "%s" + last_line
         return code % ": ".join(ecpairs) + " )"
-
-    def _print_And(self, expr):
-        PREC = precedence(expr)
-        return ' && '.join(self.parenthesize(a, PREC)
-                for a in sorted(expr.args, key=default_sort_key))
-
-    def _print_Or(self, expr):
-        PREC = precedence(expr)
-        return ' || '.join(self.parenthesize(a, PREC)
-                for a in sorted(expr.args, key=default_sort_key))
-
-    def _print_Not(self, expr):
-        PREC = precedence(expr)
-        return '!' + self.parenthesize(expr.args[0], PREC)
 
     def _print_Function(self, expr):
         if expr.func.__name__ in self.known_functions:
