+++ sympy/printing/jscode.py
@@ -188,20 +188,6 @@ class JavascriptCodePrinter(CodePrinter):
         code = "if %s" + last_line
         return code % "else if ".join(ecpairs)
 
-    def _print_And(self, expr):
-        PREC = precedence(expr)
-        return ' && '.join(self.parenthesize(a, PREC)
-                for a in sorted(expr.args, key=default_sort_key))
-
-    def _print_Or(self, expr):
-        PREC = precedence(expr)
-        return ' || '.join(self.parenthesize(a, PREC)
-                for a in sorted(expr.args, key=default_sort_key))
-
-    def _print_Not(self, expr):
-        PREC = precedence(expr)
-        return '!' + self.parenthesize(expr.args[0], PREC)
-
     def _print_Function(self, expr):
         if expr.func.__name__ in self.known_functions:
             cond_cfunc = self.known_functions[expr.func.__name__]
